chore: move preset to style file
